+++ src/Fable.Transforms/Global/Compiler.fs
@@ -2,10 +2,10 @@ namespace Fable
 
 module Literals =
     [<Literal>]
-    let VERSION = "5.0.0-alpha.11"
+    let VERSION = "5.0.0-alpha.12"
 
     [<Literal>]
-    let JS_LIBRARY_VERSION = "2.0.0-beta.2"
+    let JS_LIBRARY_VERSION = "2.0.0-beta.3"
 
 type CompilerOptionsHelper =
     static member Make