+++ src/fable-library-ts/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 2.0.0-beta.3 - 2025-03-14
+
 ### Fixed
 
 * [JS/TS] Make `nullArgCheck` report the same error message as on .NET (by @MangelMaxime)